# Optimize the statement check for a non-discarded database