--- a/ext/sqlite3/statement.c
+++ b/ext/sqlite3/statement.c
@@ -1,22 +1,12 @@
 #include <sqlite3_ruby.h>
 
 #define REQUIRE_OPEN_STMT(_ctxt) \
-  if(!_ctxt->st) \
+  if (!_ctxt->st) \
     rb_raise(rb_path2class("SQLite3::Exception"), "cannot use a closed statement");
 
-static void
-require_open_db(VALUE stmt_rb)
-{
-    VALUE closed_p = rb_funcall(
-                         rb_iv_get(stmt_rb, "@connection"),
-                         rb_intern("closed?"), 0);
-
-    if (RTEST(closed_p)) {
-        rb_raise(rb_path2class("SQLite3::Exception"),
-                 "cannot use a statement associated with a closed database");
-    }
-}
-
+#define REQUIRE_LIVE_DB(_ctxt) \
+  if (_ctxt->db->flags & SQLITE3_RB_DATABASE_DISCARDED) \
+    rb_raise(rb_path2class("SQLite3::Exception"), "cannot use a statement associated with a discarded database");
 
 VALUE cSqlite3Statement;
 
@@ -70,6 +60,11 @@ prepare(VALUE self, VALUE db, VALUE sql)
     StringValue(sql);
 
     TypedData_Get_Struct(self, sqlite3StmtRuby, &statement_type, ctx);
+
+    /* Dereferencing a pointer to the database struct will be faster than accessing it through the
+     * instance variable @connection. The struct pointer is guaranteed to be live because instance
+     * variable will keep it from being GCed. */
+    ctx->db = db_ctx;
 
 #ifdef HAVE_SQLITE3_PREPARE_V2
     status = sqlite3_prepare_v2(
@@ -135,7 +130,7 @@ step(VALUE self)
 
     TypedData_Get_Struct(self, sqlite3StmtRuby, &statement_type, ctx);
 
-    require_open_db(self);
+    REQUIRE_LIVE_DB(ctx);
     REQUIRE_OPEN_STMT(ctx);
 
     if (ctx->done_p) { return Qnil; }
@@ -232,7 +227,7 @@ bind_param(VALUE self, VALUE key, VALUE value)
 
     TypedData_Get_Struct(self, sqlite3StmtRuby, &statement_type, ctx);
 
-    require_open_db(self);
+    REQUIRE_LIVE_DB(ctx);
     REQUIRE_OPEN_STMT(ctx);
 
     switch (TYPE(key)) {
@@ -326,7 +321,7 @@ reset_bang(VALUE self)
 
     TypedData_Get_Struct(self, sqlite3StmtRuby, &statement_type, ctx);
 
-    require_open_db(self);
+    REQUIRE_LIVE_DB(ctx);
     REQUIRE_OPEN_STMT(ctx);
 
     sqlite3_reset(ctx->st);
@@ -348,7 +343,7 @@ clear_bindings_bang(VALUE self)
 
     TypedData_Get_Struct(self, sqlite3StmtRuby, &statement_type, ctx);
 
-    require_open_db(self);
+    REQUIRE_LIVE_DB(ctx);
     REQUIRE_OPEN_STMT(ctx);
 
     sqlite3_clear_bindings(ctx->st);
@@ -382,7 +377,7 @@ column_count(VALUE self)
     sqlite3StmtRubyPtr ctx;
     TypedData_Get_Struct(self, sqlite3StmtRuby, &statement_type, ctx);
 
-    require_open_db(self);
+    REQUIRE_LIVE_DB(ctx);
     REQUIRE_OPEN_STMT(ctx);
 
     return INT2NUM(sqlite3_column_count(ctx->st));
@@ -415,7 +410,7 @@ column_name(VALUE self, VALUE index)
 
     TypedData_Get_Struct(self, sqlite3StmtRuby, &statement_type, ctx);
 
-    require_open_db(self);
+    REQUIRE_LIVE_DB(ctx);
     REQUIRE_OPEN_STMT(ctx);
 
     name = sqlite3_column_name(ctx->st, (int)NUM2INT(index));
@@ -440,7 +435,7 @@ column_decltype(VALUE self, VALUE index)
 
     TypedData_Get_Struct(self, sqlite3StmtRuby, &statement_type, ctx);
 
-    require_open_db(self);
+    REQUIRE_LIVE_DB(ctx);
     REQUIRE_OPEN_STMT(ctx);
 
     name = sqlite3_column_decltype(ctx->st, (int)NUM2INT(index));
@@ -459,7 +454,7 @@ bind_parameter_count(VALUE self)
     sqlite3StmtRubyPtr ctx;
     TypedData_Get_Struct(self, sqlite3StmtRuby, &statement_type, ctx);
 
-    require_open_db(self);
+    REQUIRE_LIVE_DB(ctx);
     REQUIRE_OPEN_STMT(ctx);
 
     return INT2NUM(sqlite3_bind_parameter_count(ctx->st));
@@ -568,7 +563,7 @@ stats_as_hash(VALUE self)
     sqlite3StmtRubyPtr ctx;
     TypedData_Get_Struct(self, sqlite3StmtRuby, &statement_type, ctx);
 
-    require_open_db(self);
+    REQUIRE_LIVE_DB(ctx);
     REQUIRE_OPEN_STMT(ctx);
 
     VALUE arg = rb_hash_new();
@@ -587,7 +582,7 @@ stat_for(VALUE self, VALUE key)
     sqlite3StmtRubyPtr ctx;
     TypedData_Get_Struct(self, sqlite3StmtRuby, &statement_type, ctx);
 
-    require_open_db(self);
+    REQUIRE_LIVE_DB(ctx);
     REQUIRE_OPEN_STMT(ctx);
 
     if (SYMBOL_P(key)) {
@@ -609,7 +604,7 @@ memused(VALUE self)
     sqlite3StmtRubyPtr ctx;
     TypedData_Get_Struct(self, sqlite3StmtRuby, &statement_type, ctx);
 
-    require_open_db(self);
+    REQUIRE_LIVE_DB(ctx);
     REQUIRE_OPEN_STMT(ctx);
 
     return INT2NUM(sqlite3_stmt_status(ctx->st, SQLITE_STMTSTATUS_MEMUSED, 0));
@@ -628,7 +623,7 @@ database_name(VALUE self, VALUE index)
     sqlite3StmtRubyPtr ctx;
     TypedData_Get_Struct(self, sqlite3StmtRuby, &statement_type, ctx);
 
-    require_open_db(self);
+    REQUIRE_LIVE_DB(ctx);
     REQUIRE_OPEN_STMT(ctx);
 
     return SQLITE3_UTF8_STR_NEW2(
@@ -647,7 +642,7 @@ get_sql(VALUE self)
     sqlite3StmtRubyPtr ctx;
     TypedData_Get_Struct(self, sqlite3StmtRuby, &statement_type, ctx);
 
-    require_open_db(self);
+    REQUIRE_LIVE_DB(ctx);
     REQUIRE_OPEN_STMT(ctx);
 
     return rb_obj_freeze(SQLITE3_UTF8_STR_NEW2(sqlite3_sql(ctx->st)));
@@ -667,7 +662,7 @@ get_expanded_sql(VALUE self)
 
     TypedData_Get_Struct(self, sqlite3StmtRuby, &statement_type, ctx);
 
-    require_open_db(self);
+    REQUIRE_LIVE_DB(ctx);
     REQUIRE_OPEN_STMT(ctx);
 
     expanded_sql = sqlite3_expanded_sql(ctx->st);