--- a/ext/sqlite3/statement.h
+++ b/ext/sqlite3/statement.h
@@ -5,6 +5,7 @@
 
 struct _sqlite3StmtRuby {
     sqlite3_stmt *st;
+    sqlite3Ruby *db;
     int done_p;
 };