@flavorjones 2024-09-19
@tenderlove I'd like to merge this and cut an RC2 today, but would love your eyes on this if you get a chance.